Renamed 'benchmark/util' to 'benchmark/benchmark_util'
