+++ benchmark/metrics/large_transaction_metrics_benchmark.cpp
@@ -3,7 +3,7 @@
 #include "benchmark/benchmark.h"
 #include "metrics/metrics_thread.h"
 #include "storage/garbage_collector_thread.h"
-#include "util/data_table_benchmark_util.h"
+#include "benchmark_util/data_table_benchmark_util.h"
 
 namespace terrier {
