+++ benchmark/metrics/logging_metrics_benchmark.cpp
@@ -6,7 +6,7 @@
 #include "storage/garbage_collector_thread.h"
 #include "storage/storage_defs.h"
 #include "storage/write_ahead_log/log_manager.h"
-#include "util/data_table_benchmark_util.h"
+#include "benchmark_util/data_table_benchmark_util.h"
 
 #define LOG_FILE_NAME "/mnt/ramdisk/benchmark.txt"
