+++ benchmark/transaction/large_transaction_benchmark.cpp
@@ -1,7 +1,7 @@
 #include <vector>
 #include "benchmark/benchmark.h"
 #include "storage/garbage_collector_thread.h"
-#include "util/data_table_benchmark_util.h"
+#include "benchmark_util/data_table_benchmark_util.h"
 
 namespace terrier {
