+++ test/binder/binder_test.cpp
@@ -14,7 +14,7 @@
 #include "traffic_cop/statement.h"
 #include "transaction/deferred_action_manager.h"
 #include "transaction/transaction_manager.h"
-#include "util/data_table_benchmark_util.h"
+#include "benchmark_util/data_table_benchmark_util.h"
 #include "test_util/test_harness.h"
 
 using std::make_tuple;
